feat: introduce return & invoke
